files/4-rinkhals/opt/rinkhals/ui/demo/requirements.txt
pillow
evdev
tkinter
